chore: Remove wrong index compatibility check
The best-effort check we're doing to ensure an index
is "compatible" with a shape is wrong headed. For example, if we're
indexing a single element from an array, the output shape is []
but we're comparing it to the list of nat indices. Instead of trying
to improve the notion, I just removed it and let the indexing itself
fail if there's an incorrect index.