chore: improve iterators
TensorLib (and soon KLR) are rife with iterators. We were rolling them
ad hoc, with a lot of duplicate code. For example, an Mgrid iterator
is really just a carry-adder-like list of int iterators. A shape iterator
is just a carry-adder-like list of nat iterators. There are some weird ones
still like Slice iterators that stick around, but mostly we pulled the
generic ones into a type class.

I found a much nicer way to do carry-adder iterators, but it involved
a signature change where not all type variables are returned by all
type class functions. This means there is some annoying instruction
of the missing types in some calls. Overall, though, a massive simplification.