chore: Refactor Common.lean I
We are refactoring Common to make more utilities useful
from KLR and other importing projects. This will come
in several steps for ease of code review and comments.

This is step 1: some conversions between IntX/UIntX and ByteArray.
The goal is that we can import just the util functions we
need for a given set of types, without getting lots of names
we don't want.